+++ app/components/SearchSong.tsx
@@ -3,7 +3,7 @@ import { CheckIcon } from "@heroicons/react/20/solid";
 import { Combobox } from "@headlessui/react";
 import { BiLoaderCircle, BiSearch } from "react-icons/bi";
 import { useFetcher } from "@remix-run/react";
-import type { SpotifyTrack } from "~/utils/spotify";
+import type { SpotifyTrack } from "~/utils/spotify.server";
 import { classNames } from "~/utils/css";
 
 export const SearchSong = ({ queue_id }: { queue_id: string }) => {
